% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.configuralInvariance.R
\name{df.configuralInvariance}
\alias{df.configuralInvariance}
\title{Degrees of Freedom of Configural Invariance Model}
\usage{
df.configuralInvariance(nIndicators, nFactors, nGroups)
}
\arguments{
\item{nIndicators}{Number of indicators in model}

\item{nFactors}{Number of factors in model}

\item{nGroups}{Number of groups}
}
\value{
Degrees of freedom
}
\description{
Calculates the degrees of freedom of a model assuming configural measurement invariance (equal factor structure and equal loading patterns) across groups.
It is assumed that there are no cross-loadings, correlated item residuals, or uncorrelated factors.
}
\examples{
df.configuralInvariance(nIndicators = 6, nFactors = 2, nGroups = 2)
}
