% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.metricInvariance.R
\name{df.metricInvariance}
\alias{df.metricInvariance}
\title{Degrees of Freedom of Metric Invariance Model}
\usage{
df.metricInvariance(nIndicators, nFactors, nGroups)
}
\arguments{
\item{nIndicators}{Number of indicators in model}

\item{nFactors}{Number of factors in model}

\item{nGroups}{Number of groups}
}
\value{
Number of free parameters
}
\description{
Calculates the degrees of freedom of a model assuming metric measurement invariance (equal unstandardized factor loadings) across groups.
It is assumed that there are no cross-loadings, correlated item residuals, or uncorrelated factors.
}
\examples{
df.metricInvariance(nIndicators = 6, nFactors = 2, nGroups = 2)
}
