% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.scalarInvariance.R
\name{df.scalarInvariance}
\alias{df.scalarInvariance}
\title{Degrees of Freedom of Scalar Invariance Model}
\usage{
df.scalarInvariance(nIndicators, nFactors, nGroups)
}
\arguments{
\item{nIndicators}{Number of indicators in model}

\item{nFactors}{Number of factors in model}

\item{nGroups}{Number of groups}
}
\value{
Number of free parameters
}
\description{
Calculates the degrees of freedom of a model assuming scalar measurement invariance (equal unstandardized factor loadings, equal item intercepts) across groups.
It is assumed that there are no cross-loadings, correlated item residuals, or uncorrelated factors.
}
\examples{
df.scalarInvariance(nIndicators = 6, nFactors = 2, nGroups = 2)
}
