% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.strictInvariance.R
\name{df.strictInvariance}
\alias{df.strictInvariance}
\title{Degress of Freedom in Strict Invariance Model}
\usage{
df.strictInvariance(nIndicators, nFactors, nGroups)
}
\arguments{
\item{nIndicators}{Number of indicators in model}

\item{nFactors}{Number of factors in model}

\item{nGroups}{Number of groups}
}
\value{
Number of free parameters
}
\description{
Calculates the degrees of freedom in a model assuming strict measurement invariance (equal unstandardized factor loadings, equal item intercepts, equal item residual variances) across groups.
It is assumed that there are no cross-loadings, correlated item residuals, or uncorrelated factors.
}
\examples{
df.strictInvariance(nIndicators = 6, nFactors = 2, nGroups = 2)
}
