% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npar.configuralInvariance.R
\name{npar.configuralInvariance}
\alias{npar.configuralInvariance}
\title{Free Parameters in Configural Invariance Model}
\usage{
npar.configuralInvariance(nIndicators, nFactors, nGroups)
}
\arguments{
\item{nIndicators}{Number of indicators in model}

\item{nFactors}{Number of factors in model}

\item{nGroups}{Number of groups}
}
\value{
Number of free parameters
}
\description{
Calculates the number of free parameters in a model assuming configural measurement invariance (equal factor structure and equal loading patterns) across groups.
It is assumed that there are no cross-loadings, correlated item residuals, or uncorrelated factors.
}
\examples{
npar.configuralInvariance(nIndicators = 6, nFactors = 2, nGroups = 2)
}
