% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npar.scalarInvariance.R
\name{npar.scalarInvariance}
\alias{npar.scalarInvariance}
\title{Free Parameters in Scalar Invariance Model}
\usage{
npar.scalarInvariance(nIndicators, nFactors, nGroups)
}
\arguments{
\item{nIndicators}{Number of indicators in model}

\item{nFactors}{Number of factors in model}

\item{nGroups}{Number of groups}
}
\value{
Number of free parameters
}
\description{
Calculates the number of free parameters in a model assuming scalar measurement invariance (equal unstandardized factor loadings, equal item intercepts) across groups.
It is assumed that there are no cross-loadings, correlated item residuals, or uncorrelated factors.
}
\examples{
npar.scalarInvariance(nIndicators = 6, nFactors = 2, nGroups = 2)
}
